Nộp bài sample C, by TuanTA

+++ src/my_functions.c
@@ -1,7 +1,7 @@
 #include "my_functions.h"
  
  // Định nghĩa hàm Tinh_Tong
- float Tinh_Tong(int a, int b) {
+ int Tinh_Tong(float a, float b) {
      return a + b;
  }
